requirements.txt
gitpython==2.1.11
nltk==3.3.0